implement normal cgi handling, only by explicit full http paths though (not directories in which programs can be put).
the cgi code path is merged into the scgi codepath.  they are mostly the same, only the startup is different (scgi gets fd's over tcp and sends query differently).

specify cgi paths as:  -c /path/on/httpd {command to execute}
previously, -c was for specifying cache timeout, that is not -C.