big changes:
httpd now has -n config, for reading a config file.  it is an attrdb file, the syntax will be specified in a manual page.  the command-line options can still be used for starting the httpd, with all options, but for virtual hosting, you'll need the config file (and best stop using the flags/options).

new option -f (or config attribute ctlchan) specifies the file2chan in /chan to set up for handling commands.  currently, the only command is "reload" and reloads the database.