do not set content-type if we cannot determine it.  before this, we sent the default "application/octet-stream".  however, rfc2626 says we only SHOULD send a content-type, and if we don't know it, we MAY let client do content-detection.  that seems like a better choice.
this change came about by an article on slashdot, of all places.  for fun, read on.

a user comment had a post to a page about internet explorer security, about content-type detection among others.  IE just ignores the content-type and thinks up its own.  that's a problem when a site serves a .gif and IE interprets it as .js or .jar.  so the smart folks at microsoft have a solution:  just add option "authoritative=true" to the content-type header!  e.g.:

	Content-Type: text/plain; authoritative=true;

of course, this option has been the default since a long time and is thus completely useless.

see http://blogs.msdn.com/ie/archive/2008/07/02/ie8-security-part-v-comprehensive-protection.aspx for the details.