only style change to httpd.b, no code changed.  welcome logparse, a program for parsing log files.
logparse is still work in progress.  so far it can read a file that
describes which calculations have to be done on the log file (it
has somewhat similar operations to sql queries).  it reads through
the log file, applies all lines to the calculations/rules (called
tables), and finally outputs them on stdout.