with virtual hosting, don't look for directories named "host:port", but look for directories named "host!port".  clarify virtual host config resolving in the manual page.  fix redirections for literal ipv6 addresses.
the change to host!port is more inferno-like.  it was done after
testing httpd with ipv6:  it worked, but ip6:port directories are
ambiguous.  introducing []'s as per standard http ipv6 syntax would
be ugly.

redirection for requests with literal ipv6 addresses did not keep
the []'s in the location header.  this broke redirection for such
connections and is fixed now.